% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASin.r
\name{GWAS_integration}
\alias{GWAS_integration}
\title{A Function for GWAS integration
This package allows you to integrate a gene list of interest with large GWAS datasets.}
\usage{
GWAS_integration(GWAS_file, gene_lists, output_dir, start_pos, interval,
  interval_jumps, permutations)
}
\arguments{
\item{GWAS_file.}{GWAS .csv file containing single nucleotide polymorphisms. First five rows must be chromosome, posiition of SNP, SNP ID, P values and Q values.}

\item{gene_lists.}{Genes of interest that will be intergrated with the GWAS dataset in .csv format. First three rows must be chromosome, start position and end position of gene.}

\item{output_dir.}{Output directory to create SNP and GWAS integration output directories. Defaults to current directory}

\item{start_pos.}{Up and downstream maximum starting distance to extract proximal SNPs. Defaults to 10000.}

\item{interval.}{Amount the search distance will increase per jump. Defaults to 10000}

\item{interval_jumps.}{Amount of times the distance will increase by interval value. Defaults to 5}

\item{permutations.}{Amount of sets of random SNPs equal the observed set that will be used to create the null distribution. Defaults to 1000}
}
\value{
description The package will create two directories and output multiple datasets : Extracted_SNPS which will contain all the SNPs extracted from each gene in single column text files. The second directory will be Integration_output,and will contain a text file containing all extracted SNPs with the orginal pvalues and qvalues along side the new permuted qvalues. It will also contain histograms plotting the distributions of SNPs with qvalues < 0.05/0.1.Finally, a summary text file will contain information about each SNP file analysed, containing the file name, total SNP amount,amount of SNPs with an original qvalue < 0.05, amount of SNPs with a new qvalue < 0.05, the probability these newqvalues are by chance (0 - 100%), average number of SNPs per (permutations) set that had a new qvalue < 0.05, amount of SNPs with an original qvalue < 0.1, amount of SNPs with a new qvalue < 0.1, the probability these newqvalues are by chance (0 - 100%), average number of SNPs per (permutations) set that had a new qvalue < 0.1.
}
\description{
In summary, this package determines the significance of the distribution of SNP nominal P values within and up to (start_pos + (interval X interval_jumps))kb up and downstream to each gene of interest suspected to have an effect on GWAS phenotype. A total of (permutations) random samplings (with replacement) from the a GWAS P value data set  representing the size of each of selected SNP subsets will be generated. The q values for each SNP P value subset and all its permuted equivalents are calculated using the qvalue library in R, a package incorporated into the GWASin library. The subsequent significance level (Pperm) assigned to each of the SNP subsets is equivalent to the proportion of permutations in which at least the same number of q values < 0.05/0.1 as the SNP subset were obtained, i.e. by chance.
}
\examples{
GWAS_integration(GWAS_file = "/home/workspace/GWAS.csv", gene_lists = "/home/workspace/gene_lists/", output_dir = "/home/workspace/output/", start_pos = 1000, interval = 2000, interval_jumps = 3, permutations = 100)	
GWAS_integration(GWAS_file = "./GWAS.csv", gene_lists = "./")
}
\keyword{GWAS_file,}
\keyword{gene_lists,}
\keyword{interval,}
\keyword{interval_jumps,}
\keyword{output_dir,}
\keyword{permutations}
\keyword{start_pos,}
